new Spell and some bug fixes

+++ GameProjects/Maleficus/Assets/Scripts/Spells/ISpell.cs
@@ -12,7 +12,7 @@ public interface ISpell
 
     EPlayerID CastingPlayerID { get;  }
 
-    int HitPower { get;  }
+    float HitPower { get;  }
 
     Vector3 Direction { get; }
